`model_selector2` -> `model_selector`
